AC3 & Liberation V1.0
Added support for assassins creed 3 remastered so now it works with ac3 so it can autosplit and remove load times. Added settings for both of the games so you could decide which game u are playing please choose one or else they dont work. Also added text thing to display the current percentage and current in game time for both games. Will add a way to disable both if wanted also percentage is only useful for 100% speedruns.